services/collector/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
PyYAML==6.0.1
